requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Async
asyncio>=3.4.3
aiohttp>=3.9.0
aiofiles>=23.0.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
litellm>=1.20.0

# Hugging Face / Transformers
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0

# Vector Store
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Multimodal Processing
opencv-python>=4.8.0
pytesseract>=0.3.10
python-pptx>=0.6.21
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Video/Audio
ffmpeg-python>=0.2.0
openai-whisper>=20231117

# Utilities
loguru>=0.7.0
tenacity>=8.2.0
tiktoken>=0.5.0

# API Server (optional)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
httpx>=0.26.0
